@@ -69,7 +69,6 @@
 		"eslint-config-airbnb": "2.1.1",
 		"extract-text-webpack-plugin": "0.9.1",
 		"file-loader": "0.8.5",
-		"flow-bin": "0.20.1",
 		"less": "2.5.3",
 		"mocha": "2.3.4",
 		"nodemon": "1.8.1",
@@ -80,7 +79,7 @@
 		"webpack-dev-server": "1.14.0"
 	},
 	"engines": {
-		"node": "v0.12.x"
+		"node": "v5.2.0"
 	},
 	"scripts": {
 		"start": "NODE_ENV=${NODE_ENV:=development} DEBUG=web:* nodemon -w server server/dev.js",
