@@ -6,7 +6,7 @@ import fs from "fs";
 import path from "path";
 import debug from "debug";
 import utils from "../server/utils/utils.js";
-import {password} from "../server/utils/constants/misc.js";
+import {password, confirm, firstName, lastName, email} from "../server/utils/constants/misc.js";
 
 
 export function getControllers(...args) {
@@ -14,44 +14,45 @@ export function getControllers(...args) {
 	fs.readdirSync(path.join(__dirname, "../server/controllers")).forEach(file => {
 		if (fs.statSync(path.join(__dirname, "../server/controllers", file)).isFile()) {
 			const name = file.split(".")[0].replace(/-/g, "");
-			controllers[name] = new (require(path.join(__dirname, "../server/controllers", file)).default)(...args);
+			const Controller = require(path.join(__dirname, "../server/controllers", file)).default;
+			controllers[name] = new Controller(...args);
 		}
 	});
 	return controllers;
 }
 
-export function populate(requires, parents, children, makeDefaults) {
+export function populate(requires, data, results, makeDefaults) {
 
-	return _.times(parents.length, i => {
+	return _.times(data.length, i => {
 
 		const result = {};
 		Object.keys(requires || {}).forEach(model => {
 			if (utils.isType(requires[model], "String")) {
 				if (requires[model] === "o2o") {
-					result[model] = children[model][i];
+					result[model] = results[model][i];
 				} else if (requires[model] === "m2o") {
-					result[model] = children[model][0];
+					result[model] = results[model][0];
 				} else if (requires[model] === "o2m" || requires[model] === "m2m") {
-					result[model] = children[model];
+					result[model] = results[model];
 				} else {
 					result[model] = [];
 				}
 			} else if (utils.isType(requires[model], "Array")) {
 				if (utils.isType(requires[model][i], "Array")) {
-					result[model] = children[model].filter((item, j) => requires[model][i].indexOf(j) !== -1);
+					result[model] = results[model].filter((item, j) => requires[model][i].indexOf(j) !== -1);
 				} else if (utils.isType(requires[model][i], "Number")) {
-					result[model] = children[model][requires[model][i]];
+					result[model] = results[model][requires[model][i]];
 				} else {
 					result[model] = null;
 				}
 			} else if (utils.isType(requires[model], "Function")) {
-				result[model] = children[model].filter(requires[model](model, i));
+				result[model] = results[model].filter(requires[model](model, i));
 			} else {
 				result[model] = [];
 			}
 		});
 
-		return makeDefaults(parents[i], result, i);
+		return makeDefaults(data[i], result, i);
 	});
 
 }
@@ -59,17 +60,18 @@ export function populate(requires, parents, children, makeDefaults) {
 const controllers = getControllers(false);
 const log = debug("log:helper");
 
-export function createUser(requires, users, data) {
-	return controllers.user.create(populate(requires, users, data, (user, data, i) => Object.assign({
-		email: "ctapbiumabp" + i + "@gmail.com",
-		password: password,
-		confirm: password,
-		firstName: "Trej",
-		lastName: "Gun",
-		companyName: "Company_" + i,
-		domainName: "domain.com",
-		phoneNumber: "1234567890" + i
-	}, user)))
+export function createUser() {
+	return controllers.user.create(populate(...arguments, (user, nested, i) =>
+		Object.assign({
+			password,
+			confirm,
+			email,
+			firstName,
+			lastName,
+			companyName: "Company_" + i,
+			domainName: "domain.com",
+			phoneNumber: "1234567890" + i
+		}, user)))
 	.then(users => {
 		// log("users", users);
 		return users;
@@ -77,7 +79,8 @@ export function createUser(requires, users, data) {
 }
 
 export function cleanUp(done) {
-	return q.all(Object.keys(controllers).map(name => {
+	return q.all(Object.keys(controllers)
+	.map(name => {
 		if (name === "user") {
 			return controllers[name].find({}, {lean: false})
 			.then(users => q.allSettled(users.map(user => controllers[name].destroy(user))));
@@ -97,14 +100,14 @@ export function mockInChain(chain) {
 	let promise = q(results);
 
 	while (chain.length) {
-		promise = (next => promise.then(() => {
-			log("next", next);
+		const next = chain.shift();
+		promise = promise.then(() => { // eslint-disable-line no-loop-func
 			return module.exports["create" + next.model](next.requires, next.data || new Array(next.count).fill({}), results)
 			.then(result => {
 				results[next.model] = result;
 				return results;
 			});
-		}))(chain.shift());
+		});
 	}
 
 	return promise
@@ -138,7 +141,7 @@ export function wrapRequest(data) {
 		logIn(user, callback) {
 			callback();
 		},
-		logout () {
+		logout() {
 		}
 	}, data);
 }
